requirements.txt
groq
gradio
sentence_transformers
pdfplumber
faiss-cpu
python-dotenv
pyrebase4